robot-assistance-/Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
motor==3.7.1
pymongo==4.14.1
dnspython==2.6.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
httpx==0.27.0
